backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
python-jose[cryptography]
pydantic
python-dotenv
python-multipart
